requirements.txt
telethon
sqlalchemy
redis
python-telegram-bot==12.8
pySmartDL
html5lib
html_telegraph_poster
future
emoji
requests
psycopg2-binary
pythonping
jikanpy
feedparser
pytz
gtts
bs4
lxml
demjson
wget
bs4
telegraph
pillow
psutil
speedtest-cli>=2.0.2
googletrans
wikipedia
spamwatch
bleach
git+https://github.com/starry69/python-markdown2.git
alphabet_detector
pythonping
